refactor: rename directory and entry file
