+++ lib/util/applyTextChanges.ts
