+++ package.json
@@ -8,7 +8,7 @@
     "url": "git+https://github.com/line/ts-remove-unused"
   },
   "license": "Apache-2.0",
-  "bin": "dist/index.js",
+  "bin": "dist/main.js",
   "files": [
     "dist"
   ],
