+++ tsconfig.lib.json
@@ -14,6 +14,6 @@
     "noFallthroughCasesInSwitch": true,
     "outDir": "./node_modules/.tmp"
   },
-  "include": ["src"],
-  "exclude": ["src/**/*.test.ts"]
+  "include": ["lib"],
+  "exclude": ["lib/**/*.test.ts"]
 }
